Three new folders were added and starting files
Three new folders were added: data, old, papers.

data - For the data related to this project.
old - Contains the previous versions of this software, historic inputs
used by Micheli's thesis, a simple view of this software's architecture,
and also a report (portuguese) of what has yet to be done.
papers - For the papers which theory is relevant to this project.